libs/external/wechat-badminton-booker/requirements-pc.txt
playwright>=1.53.0
PyYAML>=6.0.2
streamlit>=1.44.0
